+++ gargoyle/compat.py
@@ -2,6 +2,12 @@ from __future__ import absolute_import, division, print_function, unicode_litera
 
 import django
 
+# Python 3.2+ includes ContextDecorator, otherwise use backport
+try:
+    from contextlib import ContextDecorator
+except ImportError:
+    from contextdecorator import ContextDecorator
+
 # Django 1.9
 
 # url(prefix, include(urls, namespace, name)) -> url(prefix, (urls, namespace, name))
@@ -9,4 +15,7 @@ if django.VERSION[:2] >= (1, 9):
     def subinclude(urls_tuple):
         return urls_tuple  # (urls, namespace, name)
 else:
-    from django.conf.urls import include as subinclude  # noqa
+    from django.conf.urls import include as subinclude
+
+
+__all__ = ['ContextDecorator', 'subinclude']