+++ gargoyle/testutils.py
@@ -7,12 +7,66 @@ gargoyle.testutils
 """
 from __future__ import absolute_import, division, print_function, unicode_literals
 
+import inspect
+import sys
+import unittest
 from functools import wraps
 
 from gargoyle import gargoyle
+from .compat import ContextDecorator
 
 
-class SwitchContextManager(object):
+class TestCaseContextDecorator(ContextDecorator):
+    """
+    ContextDecorator subclass that allows the sane decoration of TestCase classes by wrapping
+    from setUpClass to tearDownClass
+    """
+    def __call__(self, decorable):
+        if inspect.isclass(decorable):
+            if not issubclass(decorable, unittest.TestCase):
+                raise ValueError("Only supports the wrapping of unittest.TestCase classes")
+
+            klass = decorable
+
+            orig_setUpClass = klass.setUpClass
+            orig_tearDownClass = klass.tearDownClass
+
+            def setUpClass(cls):
+                self.__enter__()
+                try:
+                    if orig_setUpClass is not None:
+                        super  # Fool SuperCheckPlugin
+                        orig_setUpClass()
+                except Exception:
+                    self.__exit__(*sys.exc_info())
+                    raise
+
+            if orig_setUpClass is klass.__dict__.get('setUpClass', None):
+                # was defined on this class, state we wrap it
+                setUpClass.__wrapped__ = orig_setUpClass
+
+            def tearDownClass(cls):
+                if orig_tearDownClass is not None:
+                    super  # Fool SuperCheckPlugin
+                    orig_tearDownClass()
+                self.__exit__(None, None, None)
+
+            if orig_tearDownClass is klass.__dict__.get('tearDownClass', None):
+                # was defined on this class, state we wrap it
+                tearDownClass.__wrapped__ = orig_tearDownClass
+
+            klass.setUpClass = classmethod(setUpClass)
+            klass.tearDownClass = classmethod(tearDownClass)
+
+            return klass
+        else:
+            decorated = super(TestCaseContextDecorator, self).__call__(decorable)
+            if inspect.isfunction(decorable):
+                decorated.__wrapped__ = decorable
+            return decorated
+
+
+class SwitchContextManager(TestCaseContextDecorator):
     """
     Allows temporarily enabling or disabling a switch.
 
@@ -32,6 +86,18 @@ class SwitchContextManager(object):
     >>> def foo():
     >>>     with switches(gargoyle, my_switch_name=True):
     >>>         print(gargoyle.is_active('my_switch_name'))
+
+    Can also wrap unittest classes, which includes Django's
+    TestCase classes:
+
+    >>> @switches(my_switch_name=True)
+    ... class MyTests(TestCase):
+    ...     @classmethod
+    ...     def setUpTestData(cls):
+    ...         # my_switch_name is True here
+    ...
+    ...     def test_foo(self):
+    ...          # ... and here
     """
     def __init__(self, gargoyle=gargoyle, **keys):
         self.gargoyle = gargoyle